More user-friendly code editor

+++ mainapp/my_thread.py
@@ -51,12 +51,13 @@ class SwapCleaner(threading.Thread):
                 self.sessions.remove(i)
                 del self.status[i]
 
-            files = os.listdir(self.swap_dir)
-            for f in files:
-                abs_f = os.path.join(self.swap_dir, f)
-                if os.path.exists(abs_f) and (not os.path.isfile(abs_f)):
-                    if f not in self.status:
-                        shutil.rmtree(abs_f)
+            if os.path.exists(self.swap_dir):
+                files = os.listdir(self.swap_dir)
+                for f in files:
+                    abs_f = os.path.join(self.swap_dir, f)
+                    if os.path.exists(abs_f) and (not os.path.isfile(abs_f)):
+                        if f not in self.status:
+                            shutil.rmtree(abs_f)
             time.sleep(self.timeout)
